Move the metadata files to burdock.metadata root
