+++ evaluation/exploration.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 import os
 import copy
-from burdock.query.sql.metadata.metadata import *
+from burdock.metadata.collection import *
 
 class Exploration:
     def __init__(self, dataset_size = 3):
