--- sdk/burdock/metadata/__init__.py
@@ -1,6 +1,0 @@
-from burdock.query.sql.metadata.metadata import (Collection, Table, String,
-                                                 Boolean, DateTime, Int,
-                                                 Float, Unknown)
-
-__all__ = ["Collection", "Table", "String", "Boolean", "DateTime",
-           "Int", "Float", "Unknown"]
