+++ sdk/burdock/query/sql/__init__.py
@@ -1,4 +1,4 @@
-from .metadata.config_file import MetadataLoader
+from burdock.metadata.config_file import MetadataLoader
 from .ast.parse import QueryParser
 from .ast.validate import Validate
 from .private.rewrite import Rewriter
