+++ tests/test_local_analyst.py
@@ -2,8 +2,9 @@
 import sklearn.datasets
 import pandas as pd
 
-from burdock.query.sql import DataFrameReader, MetadataLoader, execute_private_query
-from burdock.metadata import Collection, Table, Float
+
+from burdock.query.sql import DataFrameReader, execute_private_query
+from burdock.metadata.collection import Collection, Table, Float
 
 def test_sklearn_query():
    sklearn_dataset = sklearn.datasets.load_iris()
